Updates to overview layout and Navbar / Included `.module` in the actual file name
