+++ src/app/examples/overview/overview.component.scss
@@ -3,6 +3,7 @@ ngb-progressbar {
     margin-top: 5rem;
 }
 
+
 // ## Variable Overrides ## //
 $text-dark-grey: rgb(53, 53, 53);
 $text-light: rgb(255, 255, 255);
